controllers/index: Move data loading into the route
