+++ app/templates/index-loading.hbs
@@ -1,0 +1,1 @@
+<Notification @text={{if (media "isMobile") "Loading..." "Loading task and competitor list..."}}/>
