+++ app/templates/index.hbs
@@ -1,9 +1,5 @@
-{{#if loadDataTask.isRunning}}
-  <Notification @text={{if (media "isMobile") "Loading..." "Loading task and competitor list..."}}/>
-{{else}}
-  {{#if (and hasDeviceFilter this.scoring.task)}}
-    <div class="scores-wrapper" style="height: 30%; overflow-y: scroll;">
-      <ScoreTable @task={{this.scoring.task}}/>
-    </div>
-  {{/if}}
+{{#if (and hasDeviceFilter this.scoring.task)}}
+  <div class="scores-wrapper" style="height: 30%; overflow-y: scroll;">
+    <ScoreTable @task={{this.scoring.task}}/>
+  </div>
 {{/if}}
